pyproject.toml
[project]
name = "pdf-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "phuongnh52", email = "v.phuongnh52@vinit.tech" }
]
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=23.1.0",
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.25.1",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.38.19",
    "captcha>=0.7.1",
    "chainlit>=2.5.5",
    "duckduckgo-search>=8.0.2",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "nest-asyncio>=1.6.0",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "openai-agents>=0.0.15",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pymeshlab>=2023.12.post3",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "scipy>=1.15.3",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
    "streamlit-cookies-manager>=0.2.0",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.2",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "uvicorn>=0.34.2",
]

[project.scripts]
pdf-search = "pdf_search:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_search"]
